add controllers and chain route
